+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/PlaylistPanel.java
@@ -21,8 +21,7 @@ package is.codion.framework.demos.chinook.ui;
 import is.codion.framework.demos.chinook.domain.Chinook.PlaylistTrack;
 import is.codion.swing.framework.model.SwingEntityModel;
 import is.codion.swing.framework.ui.EntityPanel;
-
-import static is.codion.swing.framework.ui.TabbedDetailLayout.splitPaneResizeWeight;
+import is.codion.swing.framework.ui.TabbedDetailLayout;
 
 public final class PlaylistPanel extends EntityPanel {
 
@@ -30,7 +29,9 @@ public final class PlaylistPanel extends EntityPanel {
     super(playlistModel,
             new PlaylistEditPanel(playlistModel.editModel()),
             new PlaylistTablePanel(playlistModel.tableModel()),
-            config -> config.detailLayout(splitPaneResizeWeight(0.25)));
+            config -> config.detailLayout(TabbedDetailLayout.builder()
+                    .splitPaneResizeWeight(0.25)
+                    .build()));
 
     SwingEntityModel playlistTrackModel = playlistModel.detailModel(PlaylistTrack.TYPE);
     EntityPanel playlistTrackPanel = new EntityPanel(playlistTrackModel,