+++ changelog.md
@@ -8,6 +8,7 @@ Codion Change Log
 - EntityPanel.containsDetailPanel() removed.
 - EntityPanel.detailPanel(entityType) refactored, error message improved.
 - EntityPanel bug fixed, editControlTablePanel was not created in case of no table panel, renamed mainPanel, tests added.
+- TabbedDetailLayout, static factory methods removed.
 
 ## 0.17.33
 ### is.codion.common.core